requirements.txt
google-adk>=1.17.0
google-cloud-firestore>=2.19.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
playwright>=1.52.0
termcolor>=3.1.0
rich
qrcode>=7.4
Pillow>=10.0.0
numpy>=1.24.0
cryptography>=41.0.0
